# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
      "Assertion": {
        "properties": {
          "asserted_by": {
            "title": "Asserted By",
            "type": "string"
          },
          "citation_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Citation Ids",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "field": {
            "title": "Field",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "sensitivity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Sensitivity"
              },
              {
                "type": "null"
              }
            ]
          },
          "space_id": {
            "format": "uuid",
            "title": "Space Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/AssertionStatus"
          },
          "subject_id": {
            "format": "uuid",
            "title": "Subject Id",
            "type": "string"
          },
          "subject_type": {
            "$ref": "#/components/schemas/SubjectType"
          },
          "supersedes_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supersedes Id"
          },
          "value": {
            "$ref": "#/components/schemas/JsonValue"
          }
        },
        "required": [
          "id",
          "space_id",
          "subject_type",
          "subject_id",
          "field",
          "value",
          "status",
          "asserted_by",
          "citation_ids",
          "supersedes_id",
          "sensitivity",
          "created_at"
        ],
        "title": "Assertion",
        "type": "object"
      },
      "AssertionCitationLink": {
        "additionalProperties": false,
        "properties": {
          "citation_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 50,
            "minItems": 1,
            "title": "Citation Ids",
            "type": "array"
          }
        },
        "required": [
          "citation_ids"
        ],
        "title": "AssertionCitationLink",
        "type": "object"
      },
      "AssertionCreate": {
        "additionalProperties": false,
        "properties": {
          "citation_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 50,
            "title": "Citation Ids",
            "type": "array"
          },
          "field": {
            "pattern": "^[a-z][a-z0-9_.]{0,99}$",
            "title": "Field",
            "type": "string"
          },
          "sensitivity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Sensitivity"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/AssertionStatus",
            "default": "suggested"
          },
          "subject_id": {
            "format": "uuid",
            "title": "Subject Id",
            "type": "string"
          },
          "subject_type": {
            "$ref": "#/components/schemas/SubjectType"
          },
          "value": {
            "$ref": "#/components/schemas/JsonValue"
          }
        },
        "required": [
          "subject_type",
          "subject_id",
          "field",
          "value"
        ],
        "title": "AssertionCreate",
        "type": "object"
      },
      "AssertionStatus": {
        "enum": [
          "suggested",
          "accepted",
          "disputed",
          "retracted"
        ],
        "title": "AssertionStatus",
        "type": "string"
      },
      "AssertionStatusChange": {
        "additionalProperties": false,
        "properties": {
          "citation_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 50,
            "title": "Citation Ids",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/AssertionStatus"
          }
        },
        "required": [
          "status"
        ],
        "title": "AssertionStatusChange",
        "type": "object"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "JsonValue": {},
...
      "Sensitivity": {
        "enum": [
          "religion",
          "health",
          "genetic",
          "ethnicity",
          "sexual",
          "political"
        ],
        "title": "Sensitivity",
        "type": "string"
      },
...
      "SubjectType": {
        "enum": [
          "person",
          "event",
          "relationship",
          "place"
        ],
        "title": "SubjectType",
        "type": "string"
      },
...
    }
  },
  "info": {
    "summary": "Evidence-based family trees and family memory. AGPL-3.0-only.",
    "title": "family-history API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/assertions/{assertion_id}/citations": {
      "post": {
        "description": "Link citations to an assertion (a new version with the same status).",
        "operationId": "link_citations_v1_assertions__assertion_id__citations_post",
        "parameters": [
          {
            "in": "path",
            "name": "assertion_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Assertion Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssertionCitationLink"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Assertion"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "summary": "Link Citations",
        "tags": [
          "assertions"
        ]
      }
    },
    "/v1/assertions/{assertion_id}/status": {
      "post": {
        "description": "Write a new version with the new status. Editors decide; accepting needs a citation.",
        "operationId": "change_status_v1_assertions__assertion_id__status_post",
        "parameters": [
          {
            "in": "path",
            "name": "assertion_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Assertion Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssertionStatusChange"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Assertion"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "summary": "Change Status",
        "tags": [
          "assertions"
        ]
      }
    },
...
    "/v1/spaces/{space_id}/assertions": {
      "get": {
        "description": "Assertions about one subject. By default only the current version of each.",
        "operationId": "list_assertions_v1_spaces__space_id__assertions_get",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "subject_type",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/SubjectType"
            }
          },
          {
            "in": "query",
            "name": "subject_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Subject Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "include_history",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include History",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Assertion"
                  },
                  "title": "Response List Assertions V1 Spaces  Space Id  Assertions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "summary": "List Assertions",
        "tags": [
          "assertions"
        ]
      },
      "post": {
        "description": "Contributors add assertions. Accepting one needs a citation; AI output only suggests.",
        "operationId": "create_assertion_v1_spaces__space_id__assertions_post",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssertionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Assertion"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "summary": "Create Assertion",
        "tags": [
          "assertions"
        ]
      }
    },
...
  }
}